[#5ekwhf]
Two groups of IndicatorServiceTest due to dependencies on two different mockIndicatorList() results.
